+++ example/sample-project/bar.js
@@ -4,3 +4,4 @@ let x2 = [,,,2, 3]
 let template = `${xx} x`
 let t2 = `12
 3`;
+let badTemplate = `bad template`;
