+++ tsconfig.json
@@ -7,7 +7,7 @@
     "strict": true, /* Enable all strict type-checking options. */
     "outDir": "build",
     "moduleResolution": "node",
-    "allowSyntheticDefaultImports": true ,
+    "allowSyntheticDefaultImports": true
   },
   "include": ["src"]
 }
